requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
matplotlib==3.10.7
mcp==1.19.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
